requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
exceptiongroup==1.1.1
frozendict==2.3.7
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gspread==5.8.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
lxml==4.9.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
webdriver-manager==3.8.6
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.18
